Add files via upload
I have created the website for PK Jewellers using HTML and CSS. In this 4 webpages are created. In 2 Pages I have added logo. and for all 4 pages, I have added background images.
In my first webpage, I created 5 inline-block elements which mean I have provided 4 categories for jwellery Gold, Silver, Dimond, and others. For this, I have added 4 images related to the categories on the first webpage. also on the first web page, I have interlinked 2 web pages that is a login page and one more jwellery page. suppose if we click on gold it will take you to the gold jwellery page. 
if you click on the login then the login form will open and it will give you the option like sign-in with Google, sign-up, and sign-in.  and I provided an option for if you do not have an account then click on Sign-up.
After clicking on the Sign-up option one form will open and for that form, I have provided the form validation option.
In this project, I used different CSS properties to set the background color, and image size, placing elements at a certain position on the browser and styling the webpage.
Also, I have used the Div tag to create a document structure.
I have used the Internal and External CSS types for creating this website